requirements.txt
flask
gunicorn
groq
sentence-transformers
faiss-cpu
python-dotenv
pypdf2
numpy
requests
